backend/requirements.txt
Flask
flask-cors
mysql-connector-python
SQLAlchemy
PyMySQL
python-dotenv
